renaming multiple files, folders and components to add lumina as prefix
