+++ apps/lumina-core/src/components/editor/editor.tsx
@@ -2,7 +2,7 @@
 
 import styles from "@/components/editor/editor.module.scss";
 import cx from "classnames";
-import { SidebarEditor } from "../sidebar/sidebar";
+import { LuminaSidebarEditor } from "../lumina-sidebar/lumina-sidebar";
 import { useCallback, useState } from "react";
 
 type Props = {
@@ -19,7 +19,7 @@ export const Editor = ({ children }: Props) => {
   return (
     <div
     className={cx(styles.editorContainer, { [String(styles.open)]: isBarOpen, })}>
-      <SidebarEditor handleToggler={handleMenuToggler} isBarOpen={isBarOpen}/>
+      <LuminaSidebarEditor handleToggler={handleMenuToggler} isBarOpen={isBarOpen}/>
       <div>{children}</div>
     </div>
   );
