+++ apps/lumina-core/src/components/lumina-action-buttons/delete/lumina-delete-component.tsx
@@ -1,5 +1,5 @@
-import { Button } from "@/components/button/buttons";
-import { DeleteModal } from "@/components/lumina-modals/delete/delete-modal";
+import { LuminaButton } from "@/components/lumina-button/lumina-button";
+import { LuminaDeleteModal } from "@/components/lumina-modals/delete/lumina-delete-modal";
 import { useLuminaContext } from "@/context/contextProvider";
 import { useCallback, useState } from "react";
 
@@ -30,13 +30,13 @@ export const DeleteComponentButton = ({ id }: TProps) => {
   )
 
   return (
-    <><Button
+    <><LuminaButton
       color="danger"
       onClick={handleToggleDeleteModal}
       round
       iconLeft="lumina-delete"
     />
-      <DeleteModal showModalDelete={showModalDelete}
+      <LuminaDeleteModal showModalDelete={showModalDelete}
         handleCloseModal={handleToggleDeleteModal}
         handleOnClickDelete={handleOnClickDelete}
       />
