+++ apps/lumina-core/src/components/lumina-action-buttons/edit/lumina-edit-component.tsx
@@ -1,5 +1,5 @@
-import { Button } from "@/components/button/buttons";
-import { EditModal } from "@/components/lumina-modals/edit/edit-modal";
+import { LuminaButton } from "@/components/lumina-button/lumina-button";
+import { LuminaEditModal } from "@/components/lumina-modals/edit/lumina-edit-modal";
 import { useLuminaContext } from "@/context/contextProvider";
 import { IComponentProps } from "@/data/data";
 import { TConfig } from "@/models/showEditModel";
@@ -54,13 +54,13 @@ export const EditComponentButton = ({ id, data, onUpdate, config }: TProps) => {
 
   return (
     <>
-      <Button
+      <LuminaButton
         onClick={handleToggleEditModal}
         round
         iconLeft="lumina-edit"
       />
 
-      <EditModal showModalEdit={showModalEdit}
+      <LuminaEditModal showModalEdit={showModalEdit}
         handleCloseModal={handleToggleEditModal}
         handleOnClickSaveData={handleOnClickSaveData}
         handleOnChangeInput={handleOnChangeInput}
