+++ apps/lumina-core/src/components/lumina-action-buttons/menu/lumina-menu-component.tsx
@@ -1,8 +1,8 @@
-import { Button } from "@/components/button/buttons"
+import { LuminaButton } from "@/components/lumina-button/lumina-button"
 
 export const MenuComponentButton = () => {
   return (
-    <Button
+    <LuminaButton
       color="secondary"
       iconLeft={"lumina-menu"}
     />
