+++ apps/lumina-core/src/components/lumina-action-buttons/visible/lumina-visible-component.tsx
@@ -1,9 +1,9 @@
-import { Button } from "@/components/button/buttons"
+import { LuminaButton } from "@/components/lumina-button/lumina-button"
 
 export const VisibleComponentButton = () => {
 
   return (
-    <Button
+    <LuminaButton
       color="secondary"
       iconLeft={"lumina-visible"}
     />
