+++ apps/lumina-core/src/components/lumina-sidebar/lumina-sidebar.module.scss
