+++ apps/lumina-core/src/components/lumina-tabs/pages/lumina-page.module.scss
